+++ packages/frontend/src/app/(auth)/login/page.tsx
@@ -1,0 +1,73 @@
+import { login } from '../actions'; // Your login and signup action handlers
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Label } from '@/components/ui/label';
+import { Input } from '@/components/ui/input';
+import Link from 'next/link';
+import { GitHubLogoIcon } from '@radix-ui/react-icons';
+
+export default function LoginPage() {
+  return (
+    <div className="flex items-center justify-center min-h-screen">
+      <Card className="mx-auto max-w-sm">
+        <CardHeader>
+          <CardTitle className="text-2xl">Login</CardTitle>
+          <CardDescription>
+            Enter your email below to log in to your account
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form method="post">
+            <div className="grid gap-4">
+              <div className="grid gap-2">
+                <Label htmlFor="email">Email</Label>
+                <Input
+                  id="email"
+                  name="email"
+                  type="email"
+                  placeholder="m@example.com"
+                  required
+                />
+              </div>
+              <div className="grid gap-2">
+                <div className="flex items-center">
+                  <Label htmlFor="password">Password</Label>
+                  <Link href="#" className="ml-auto inline-block text-sm underline">
+                    Forgot your password?
+                  </Link>
+                </div>
+                <Input
+                  id="password"
+                  name="password"
+                  type="password"
+                  required
+                />
+              </div>
+              <Button
+                type="submit"
+                formAction={login}
+                className="w-full"
+              >
+                Log in
+              </Button>
+              <Button
+                type="submit"
+                variant="outline"
+                className="w-full"
+              >
+                <GitHubLogoIcon />
+                Continue with GitHub
+              </Button>
+            </div>
+            <div className="mt-4 text-center text-sm">
+              Don&apos;t have an account?{" "}
+              <Link href="/signup" className="underline">
+                Sign up
+              </Link>
+            </div>
+          </form>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
