+++ packages/frontend/src/app/(auth)/signup/page.tsx
@@ -1,0 +1,68 @@
+import { signup } from '../actions';
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Label } from '@/components/ui/label';
+import { Input } from '@/components/ui/input';
+import Link from 'next/link';
+import { GitHubLogoIcon } from '@radix-ui/react-icons';
+
+export default function SignupPage() {
+  return (
+    <div className="flex items-center justify-center min-h-screen">
+      <Card className="mx-auto max-w-sm">
+        <CardHeader>
+          <CardTitle className="text-2xl">Sign Up</CardTitle>
+          <CardDescription>
+            Enter your email below to create a new account
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form method="post">
+            <div className="grid gap-4">
+              <div className="grid gap-2">
+                <Label htmlFor="email">Email</Label>
+                <Input
+                  id="email"
+                  name="email"
+                  type="email"
+                  placeholder="m@example.com"
+                  required
+                />
+              </div>
+              <div className="grid gap-2">
+                <Label htmlFor="password">Password</Label>
+                <Input
+                  id="password"
+                  name="password"
+                  type="password"
+                  required
+                />
+              </div>
+              <Button
+                type="submit"
+                formAction={signup}
+                className="w-full"
+              >
+                Sign up
+              </Button>
+              <Button
+                type="submit"
+                variant="outline"
+                className="w-full"
+              >
+                <GitHubLogoIcon />
+                Continue with GitHub
+              </Button>
+            </div>
+            <div className="mt-4 text-center text-sm">
+              Already have an account?{" "}
+              <Link href="/login" className="underline">
+                Log in
+              </Link>
+            </div>
+          </form>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
